app/requirements.txt
fastapi==0.112.2
fastapi-cli==0.0.5
faiss-cpu
langchain
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langdetect
langchainhub
pandas
pycparser
pydantic
pydantic-extra-types
pydantic-settings
pydantic_core
python-dotenv
unstructured
unstructured-client
uvicorn